requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
flask-login==0.6.3
python-dateutil==2.9.0.post0
gunicorn==21.2.0
